wgpu-core: Move `Global` to new `global` module.
