+++ wgpu-core/src/error.rs
@@ -1,10 +1,7 @@
 use core::fmt;
 use std::error::Error;
 
-use crate::{
-    gfx_select,
-    hub::{Global, IdentityManagerFactory},
-};
+use crate::{gfx_select, global::Global, hub::IdentityManagerFactory};
 
 pub struct ErrorFormatter<'a> {
     writer: &'a mut dyn fmt::Write,
