+++ wgpu-core/src/hub.rs
@@ -153,6 +153,7 @@ use crate::{
     binding_model::{BindGroup, BindGroupLayout, PipelineLayout},
     command::{CommandBuffer, RenderBundle},
     device::Device,
+    global::Global,
     id,
     instance::{Adapter, HalSurface, Instance, Surface},
     pipeline::{ComputePipeline, RenderPipeline, ShaderModule},
@@ -243,7 +244,7 @@ impl IdentityManager {
 
 /// An entry in a `Storage::map` table.
 #[derive(Debug)]
-enum Element<T> {
+pub(crate) enum Element<T> {
     /// There are no live ids with this index.
     Vacant,
 
@@ -282,7 +283,7 @@ pub(crate) struct InvalidId;
 /// that keeps the index values dense and close to zero.
 #[derive(Debug)]
 pub struct Storage<T, I: id::TypedId> {
-    map: Vec<Element<T>>,
+    pub(crate) map: Vec<Element<T>>,
     kind: &'static str,
     _phantom: PhantomData<I>,
 }
@@ -454,7 +455,7 @@ impl<T, I: id::TypedId> Storage<T, I> {
         self.map.len()
     }
 
-    fn generate_report(&self) -> StorageReport {
+    pub(crate) fn generate_report(&self) -> StorageReport {
         let mut report = StorageReport {
             element_size: mem::size_of::<T>(),
             ..Default::default()
@@ -755,7 +756,7 @@ pub trait Resource {
 #[derive(Debug)]
 pub struct Registry<T: Resource, I: id::TypedId, F: IdentityHandlerFactory<I>> {
     identity: F::Filter,
-    data: RwLock<Storage<T, I>>,
+    pub(crate) data: RwLock<Storage<T, I>>,
     backend: Backend,
 }
 
@@ -772,7 +773,7 @@ impl<T: Resource, I: id::TypedId, F: IdentityHandlerFactory<I>> Registry<T, I, F
         }
     }
 
-    fn without_backend(factory: &F, kind: &'static str) -> Self {
+    pub(crate) fn without_backend(factory: &F, kind: &'static str) -> Self {
         Self {
             identity: factory.spawn(),
             data: RwLock::new(Storage {
@@ -1068,7 +1069,11 @@ impl<A: HalApi, F: GlobalIdentityHandlerFactory> Hub<A, F> {
     //TODO: instead of having a hacky `with_adapters` parameter,
     // we should have `clear_device(device_id)` that specifically destroys
     // everything related to a logical device.
-    fn clear(&self, surface_guard: &mut Storage<Surface, id::SurfaceId>, with_adapters: bool) {
+    pub(crate) fn clear(
+        &self,
+        surface_guard: &mut Storage<Surface, id::SurfaceId>,
+        with_adapters: bool,
+    ) {
         use crate::resource::TextureInner;
         use hal::{Device as _, Surface as _};
 
@@ -1258,19 +1263,19 @@ impl<A: HalApi, F: GlobalIdentityHandlerFactory> Hub<A, F> {
 
 pub struct Hubs<F: GlobalIdentityHandlerFactory> {
     #[cfg(all(feature = "vulkan", not(target_arch = "wasm32")))]
-    vulkan: Hub<hal::api::Vulkan, F>,
+    pub(crate) vulkan: Hub<hal::api::Vulkan, F>,
     #[cfg(all(feature = "metal", any(target_os = "macos", target_os = "ios")))]
-    metal: Hub<hal::api::Metal, F>,
+    pub(crate) metal: Hub<hal::api::Metal, F>,
     #[cfg(all(feature = "dx12", windows))]
-    dx12: Hub<hal::api::Dx12, F>,
+    pub(crate) dx12: Hub<hal::api::Dx12, F>,
     #[cfg(all(feature = "dx11", windows))]
-    dx11: Hub<hal::api::Dx11, F>,
+    pub(crate) dx11: Hub<hal::api::Dx11, F>,
     #[cfg(feature = "gles")]
-    gl: Hub<hal::api::Gles, F>,
+    pub(crate) gl: Hub<hal::api::Gles, F>,
 }
 
 impl<F: GlobalIdentityHandlerFactory> Hubs<F> {
-    fn new(factory: &F) -> Self {
+    pub(crate) fn new(factory: &F) -> Self {
         Self {
             #[cfg(all(feature = "vulkan", not(target_arch = "wasm32")))]
             vulkan: Hub::new(factory),
@@ -1282,153 +1287,6 @@ impl<F: GlobalIdentityHandlerFactory> Hubs<F> {
             dx11: Hub::new(factory),
             #[cfg(feature = "gles")]
             gl: Hub::new(factory),
-        }
-    }
-}
-
-#[derive(Debug)]
-pub struct GlobalReport {
-    pub surfaces: StorageReport,
-    #[cfg(all(feature = "vulkan", not(target_arch = "wasm32")))]
-    pub vulkan: Option<HubReport>,
-    #[cfg(all(feature = "metal", any(target_os = "macos", target_os = "ios")))]
-    pub metal: Option<HubReport>,
-    #[cfg(all(feature = "dx12", windows))]
-    pub dx12: Option<HubReport>,
-    #[cfg(all(feature = "dx11", windows))]
-    pub dx11: Option<HubReport>,
-    #[cfg(feature = "gles")]
-    pub gl: Option<HubReport>,
-}
-
-pub struct Global<G: GlobalIdentityHandlerFactory> {
-    pub instance: Instance,
-    pub surfaces: Registry<Surface, id::SurfaceId, G>,
-    hubs: Hubs<G>,
-}
-
-impl<G: GlobalIdentityHandlerFactory> Global<G> {
-    pub fn new(name: &str, factory: G, instance_desc: wgt::InstanceDescriptor) -> Self {
-        profiling::scope!("Global::new");
-        Self {
-            instance: Instance::new(name, instance_desc),
-            surfaces: Registry::without_backend(&factory, "Surface"),
-            hubs: Hubs::new(&factory),
-        }
-    }
-
-    /// # Safety
-    ///
-    /// Refer to the creation of wgpu-hal Instance for every backend.
-    pub unsafe fn from_hal_instance<A: HalApi>(
-        name: &str,
-        factory: G,
-        hal_instance: A::Instance,
-    ) -> Self {
-        profiling::scope!("Global::new");
-        Self {
-            instance: A::create_instance_from_hal(name, hal_instance),
-            surfaces: Registry::without_backend(&factory, "Surface"),
-            hubs: Hubs::new(&factory),
-        }
-    }
-
-    /// # Safety
-    ///
-    /// - The raw instance handle returned must not be manually destroyed.
-    pub unsafe fn instance_as_hal<A: HalApi>(&self) -> Option<&A::Instance> {
-        A::instance_as_hal(&self.instance)
-    }
-
-    /// # Safety
-    ///
-    /// - The raw handles obtained from the Instance must not be manually destroyed
-    pub unsafe fn from_instance(factory: G, instance: Instance) -> Self {
-        profiling::scope!("Global::new");
-        Self {
-            instance,
-            surfaces: Registry::without_backend(&factory, "Surface"),
-            hubs: Hubs::new(&factory),
-        }
-    }
-
-    pub fn clear_backend<A: HalApi>(&self, _dummy: ()) {
-        let mut surface_guard = self.surfaces.data.write();
-        let hub = A::hub(self);
-        // this is used for tests, which keep the adapter
-        hub.clear(&mut surface_guard, false);
-    }
-
-    pub fn generate_report(&self) -> GlobalReport {
-        GlobalReport {
-            surfaces: self.surfaces.data.read().generate_report(),
-            #[cfg(all(feature = "vulkan", not(target_arch = "wasm32")))]
-            vulkan: if self.instance.vulkan.is_some() {
-                Some(self.hubs.vulkan.generate_report())
-            } else {
-                None
-            },
-            #[cfg(all(feature = "metal", any(target_os = "macos", target_os = "ios")))]
-            metal: if self.instance.metal.is_some() {
-                Some(self.hubs.metal.generate_report())
-            } else {
-                None
-            },
-            #[cfg(all(feature = "dx12", windows))]
-            dx12: if self.instance.dx12.is_some() {
-                Some(self.hubs.dx12.generate_report())
-            } else {
-                None
-            },
-            #[cfg(all(feature = "dx11", windows))]
-            dx11: if self.instance.dx11.is_some() {
-                Some(self.hubs.dx11.generate_report())
-            } else {
-                None
-            },
-            #[cfg(feature = "gles")]
-            gl: if self.instance.gl.is_some() {
-                Some(self.hubs.gl.generate_report())
-            } else {
-                None
-            },
-        }
-    }
-}
-
-impl<G: GlobalIdentityHandlerFactory> Drop for Global<G> {
-    fn drop(&mut self) {
-        profiling::scope!("Global::drop");
-        log::info!("Dropping Global");
-        let mut surface_guard = self.surfaces.data.write();
-
-        // destroy hubs before the instance gets dropped
-        #[cfg(all(feature = "vulkan", not(target_arch = "wasm32")))]
-        {
-            self.hubs.vulkan.clear(&mut surface_guard, true);
-        }
-        #[cfg(all(feature = "metal", any(target_os = "macos", target_os = "ios")))]
-        {
-            self.hubs.metal.clear(&mut surface_guard, true);
-        }
-        #[cfg(all(feature = "dx12", windows))]
-        {
-            self.hubs.dx12.clear(&mut surface_guard, true);
-        }
-        #[cfg(all(feature = "dx11", windows))]
-        {
-            self.hubs.dx11.clear(&mut surface_guard, true);
-        }
-        #[cfg(feature = "gles")]
-        {
-            self.hubs.gl.clear(&mut surface_guard, true);
-        }
-
-        // destroy surfaces
-        for element in surface_guard.map.drain(..) {
-            if let Element::Occupied(surface, _) = element {
-                self.instance.destroy_surface(surface);
-            }
         }
     }
 }
@@ -1580,12 +1438,6 @@ impl HalApi for hal::api::Gles {
     fn get_surface_mut(surface: &mut Surface) -> Option<&mut HalSurface<Self>> {
         surface.gl.as_mut()
     }
-}
-
-#[cfg(test)]
-fn _test_send_sync(global: &Global<IdentityManagerFactory>) {
-    fn test_internal<T: Send + Sync>(_: T) {}
-    test_internal(global)
 }
 
 #[test]
