+++ wgpu-core/src/id.rs
@@ -42,7 +42,7 @@ type Dummy = hal::api::Empty;
 /// `X<Empty>` type with the resource type `X<A>`, for some specific backend
 /// `A`.
 ///
-/// [`Global`]: crate::hub::Global
+/// [`Global`]: crate::global::Global
 /// [`Hub`]: crate::hub::Hub
 /// [`Hub<A>`]: crate::hub::Hub
 /// [`Storage`]: crate::hub::Storage
